Move old codegen into a separate module
